Uzume child init

+++ child-test.sh
@@ -32,7 +32,11 @@ do
     echo "Pour $fichier"
     rm $pwd/$fichier
     #cp $pwd/child/$child/$fichier $pwd/$fichier
-    ln $pwd/child/$child/$fichier $pwd/$fichier
+    if [ -f "$pwd/child/$child/$fichier" ] ; then
+        ln $pwd/child/$child/$fichier $pwd/$fichier
+    else 
+        ln $pwd/child/default/$fichier $pwd/$fichier
+    fi
     echo "RewriteCond %{DOCUMENT_ROOT}/child/%{HTTP_HOST}/$fichier -f" >> $htaccess
     echo "RewriteRule ^$fichier$ child/%{HTTP_HOST}/$fichier [NC,QSA]" >> $htaccess
 done
